Unit testing with JUnit (#20)
Closes #11.

Unit testing is an automated system for testing individual portions of
your codebase. They allow you to setup tests that assert that something
is true. For instance:

```kotlin
import kotlin.test.Test
import kotlin.test.assertEquals

internal class DemoTest {
    @Test
    fun demoTest() {
        // Checks that 1 + 2 = 3.
        assertEquals(3, 1 + 2)
    }
}
```

This test will check that 1 + 2 = 3. If that is ever not the case, the
test will fail. We can use this to verify that logic remains remains
consistent across changes. For a more concrete example:

```kotlin
// Returns true if `this` is null.
fun <T> T?.isNull(): Boolean = this == null

@Test
fun testIsNull() {
    val x: Int? = null
    assertTrue(x.isNull())

    val y: Int? = 1
    assertFalse(x.isNull())
}
```

If `isNull` is ever changed incorrectly, the tests will fail and we will
know about it.