Add a property to detect whether an API is initialized (#35)
Occasionally you may write a class that depends on an API, but is not an
API itself so it cannot use the built-in dependencies system. This PR
adds the boolean property `API.isInit`, which returns true if the API
has been initialized.

I plan to use this as part of #26, so you don't try to use a `KiwiDrive`
without initialized `TriWheels` first.