`Polar2dDual` and `DualNum.atan()` (#44)
This PR adds the `Polar2dDual` class, which will be used with Road
Runner to calculate the position of the robot based on how far the
wheels have rotated. (`TriWheels.inverse()`, but it uses dual numbers.)

Because `Polar2dDual` requires converting to and from `Vector2dDual`, I
have also implemented `atan()` (inverse tangent) for `DualNum` and a
corresponding `atan2()` for `Vector2dDual`.

For more information on what dual numbers are, see [On dual
numbers](https://github.com/BotsBurgh/BOTSBURGH-FTC-2024-25/wiki/On-dual-numbers).