Fix `RobotConfig`'s package (#59)
The `package` line of a Kotlin file declares where an import should look
for a file, and usually corresponds to the directory structure that file
is within. This wasn't the case of `RobotConfig`, though.

While it is within the `teamcode` folder, its `package` line said it was
within `teamcode/utils`. This PR fixes that and all related imports.